Add business number method
